Wire JWT authentication middleware and protected /me endpoint
Added authentication middleware that validates JWTs and injects user identity into request extensions. Introduced protected /me endpoint and verified authorization behavior using curl.

+++ src/auth/repository.rs
@@ -11,7 +11,7 @@ pub struct AuthUserRepository {
 }
 
 struct InnerRepo {
-    next_id: u64,
+    next_id: u32,
     users_by_username: HashMap<String, AuthUser>,
 }
